// src/app/sysrev/models/Declaration.ts
export class Declaration{
    codeOffice: string="";
    office: string="";
    numeroBonSortie:string="";
    refLiquidation: string="";
    numeroEnregistrement: string="";
    numeroQuittance:string="";
    credit: string="";
    codeDeclarant: string="";
    codeEntreprise: string="";
    montant: number = 0;
    dateLiquidation:any;
    dateQuittance:any;
}