// src/app/sysrev/models/BonSortie_1.ts
import { DetailBonSortie_1 } from "./DetailBonSortie_1";


export class BonSortie_1{
    uuid: string = "";
    refbonsortie:string = "";
    dateEmission: string = "";
    dateSortie:string = "";
    refdeclaration: string = "";
    declarationUuid: string = "";
    detailBonSortieSidonyaDtos:DetailBonSortie_1[] = [];
}