// src/app/sysrev/models/DetailBonSortie.ts
export class DetailBonSortie{
    article: number = 0;
    mark_1: string = "";
    mark_2: string = "";
    nbreColis: number = 0;
    typeColis: string = "";
    natureColis: string = "";
    poids: number = 0;
    numeroConteneur: string = "";
    refManifest: string = "";
    titre:string = "";
}